+++ machine-services/machine-data-service/src/main/java/com/machine/service/data/tag/dao/mapper/entity/DataTagCategoryEntity.java
@@ -42,5 +42,6 @@ public class DataTagCategoryEntity extends BaseEntity {
      * 排序，sort值大的排序靠前
      */
     @TableField("sort")
-    private Integer sort;
+    private Long sort;
 }
+
